separated out config nodes
