+++ nodes/config/env_config.py
@@ -1,0 +1,37 @@
+import os
+
+# StructureGlobalDriversConfig,
+
+
+class gtUIEnv:
+    """
+    The Griptape Environment Config
+    Setting environment variables
+    """
+
+    @classmethod
+    def INPUT_TYPES(s):
+        return {
+            "required": {},
+            "optional": {
+                "Environment Vars": ("STRING", {"default": "ENV=", "multiline": True})
+            },
+        }
+
+    FUNCTION = "run"
+    RETURN_TYPES = ("STRING",)
+    RETURN_NAMES = ("ENVIRS",)
+    OUTPUT_NODE = True
+
+    CATEGORY = "Griptape"
+
+    def run(self, **kwargs):
+        envirs = kwargs.get("Environment Vars", "")
+        environment_vars = []
+        for envir in envirs.split("\n"):
+            if "=" in envir:
+                key, value = envir.split("=", 1)
+                if key and value:
+                    os.environ[key] = value
+                    environment_vars.append(f"{key}={value}")
+        return (environment_vars,)
